Specify uselib property to allow custom linking
Currently this addon only works with libarchive installed in the default
locations checked by the linker, even so the configure instructions do
detect if a custom PKG_CONFIG_PATH was provided.

This patch allows for the build to succeed / do the right thing by
referencing the uselib_store name used during configuration.